tested again

+++ 0x0E-structures_typedef/dog.h
@@ -16,7 +16,7 @@ typedef struct dog dog_t;
 struct dog
 {
 	char *name;
-	int age;
+	float age;
 	char *owner;
 };
 void init_dog(struct dog *d, char *name, float age, char *owner);
